lint and fmt + move workspace rbac context stuff to a separate file
